requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
Flask-WTF
Flask-Limiter
Flask-Migrate
Flask-SocketIO
Werkzeug
gunicorn
eventlet
